fix: clean error paths for users
