+++ cli/cli/main.go
@@ -6,9 +6,12 @@
 package main
 
 import (
+	"errors"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/commands"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/out"
 	"github.com/sirupsen/logrus"
 	"os"
+	"strings"
 )
 
 const (
@@ -17,6 +20,9 @@ const (
 
 	forceColors   = true
 	fullTimestamp = true
+
+	separator                      = "---"
+	removeTrailingSpecialCharacter = "\n"
 )
 
 func main() {
@@ -40,8 +46,51 @@ func main() {
 	})
 
 	if err := commands.RootCmd.Execute(); err != nil {
-		// We don't actually need to print the error because Cobra will do it for us
+		maybeCleanedError := getErrorMessageToBeDisplayedOnCli(err)
+		// cobra uses this method underneath as well to print errors
+		// so just used it directly.
+		commands.RootCmd.PrintErrln("Error: ", maybeCleanedError.Error())
 		os.Exit(errorExitCode)
 	}
 	os.Exit(successExitCode)
+}
+
+func getErrorMessageToBeDisplayedOnCli(errorWithStacktrace error) error {
+	// if we are running in the debug mode, just return the error with stack-traces back to the client
+	if logrus.GetLevel() == logrus.DebugLevel {
+		return errorWithStacktrace
+	}
+
+	// silently catch the file logger error and print it in the debug mode
+	// users should not worry about this error
+	// downside is that we may lose stack-traces during file logger failures
+	fileLogger, err := out.GetFileLogger()
+	if err != nil {
+		logrus.Debugf("Error occurred while getting the file logger %+v", err)
+	} else {
+		fileLogger.Errorln(errorWithStacktrace.Error())
+	}
+
+	errorMessage := errorWithStacktrace.Error()
+	cleanError := removeFilePathFromErrorMessage(errorMessage)
+	return cleanError
+}
+
+// this method removes the file path from the error
+func removeFilePathFromErrorMessage(errorMessage string) error {
+	errorMessageConvertedInList := strings.Split(errorMessage, separator)
+	// safe to assume that the error was not generated using stacktrace package
+	if len(errorMessageConvertedInList) == 1 {
+		return errors.New(errorMessage)
+	}
+
+	// only the even numbered elements needs to be picked.
+	var cleanErrorList []string
+	for index, line := range errorMessageConvertedInList {
+		if index%2 == 0 {
+			cleanErrorList = append(cleanErrorList, strings.Trim(line, removeTrailingSpecialCharacter))
+		}
+	}
+
+	return errors.New(strings.Join(cleanErrorList, ""))
 }
